# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Silo",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/gog/auth": {
      "post": {
        "tags": [
          "gog"
        ],
        "summary": "Connect Gog",
        "operationId": "connect_gog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GogAuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GogStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "gog"
        ],
        "summary": "Disconnect Gog",
        "operationId": "disconnect_gog",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        }
      }
    },
...
    "/api/v1/gog/import": {
      "post": {
        "tags": [
          "gog"
        ],
        "summary": "Import Gog Game",
        "operationId": "import_gog_game",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GogImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobEnqueuedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/igdb/credentials": {
      "put": {
        "tags": [
          "igdb"
        ],
        "summary": "Connect Igdb",
        "operationId": "connect_igdb",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IgdbCredentialsPut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IgdbStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "igdb"
        ],
        "summary": "Disconnect Igdb",
        "operationId": "disconnect_igdb",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        }
      }
    },
...
    "/api/v1/identification-sources/{slug}/discover-datasets": {
      "post": {
        "tags": [
          "identification"
        ],
        "summary": "Discover Datasets",
        "operationId": "discover_identification_datasets",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverDatasetsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/identification-datasets/{dataset_id}": {
      "patch": {
        "tags": [
          "identification"
        ],
        "summary": "Update Dataset",
        "operationId": "update_identification_dataset",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dataset"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/identification-datasets/{dataset_id}/refresh": {
      "post": {
        "tags": [
          "identification"
        ],
        "summary": "Refresh Dataset",
        "operationId": "refresh_identification_dataset",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobEnqueuedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Dataset": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source_slug": {
            "type": "string",
            "title": "Source Slug"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "dataset_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Version"
          },
          "refreshed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refreshed At"
          },
          "signature_count": {
            "type": "integer",
            "title": "Signature Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_slug",
          "slug",
          "name",
          "enabled",
          "dataset_version",
          "refreshed_at",
          "signature_count"
        ],
        "title": "Dataset",
        "description": "Identification dataset with its source and signature count."
      },
      "DatasetPatch": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DatasetPatch",
        "description": "Dataset settings patch."
      },
      "DiscoverDatasetsOut": {
        "properties": {
          "dataset_count": {
            "type": "integer",
            "title": "Dataset Count"
          }
        },
        "type": "object",
        "required": [
          "dataset_count"
        ],
        "title": "DiscoverDatasetsOut",
        "description": "Dataset-discovery result."
      },
...
      "GogAuthRequest": {
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "GogAuthRequest",
        "description": "Auth-code exchange request."
      },
...
      "GogImportRequest": {
        "properties": {
          "gog_id": {
            "type": "integer",
            "title": "Gog Id"
          },
          "library_id": {
            "type": "string",
            "format": "uuid",
            "title": "Library Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "gog_id",
          "library_id"
        ],
        "title": "GogImportRequest",
        "description": "Import request for one owned GOG game."
      },
      "GogStatusOut": {
        "properties": {
          "connected": {
            "type": "boolean",
            "title": "Connected"
          }
        },
        "type": "object",
        "required": [
          "connected"
        ],
        "title": "GogStatusOut",
        "description": "GOG connection status."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IgdbCredentialsPut": {
        "properties": {
          "client_id": {
            "type": "string",
            "minLength": 1,
            "title": "Client Id"
          },
          "client_secret": {
            "type": "string",
            "minLength": 1,
            "title": "Client Secret"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "client_id",
          "client_secret"
        ],
        "title": "IgdbCredentialsPut",
        "description": "Twitch developer-app credentials for IGDB."
      },
      "IgdbStatusOut": {
        "properties": {
          "connected": {
            "type": "boolean",
            "title": "Connected"
          }
        },
        "type": "object",
        "required": [
          "connected"
        ],
        "title": "IgdbStatusOut",
        "description": "IGDB connection status."
      },
...
      "JobEnqueuedOut": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "JobEnqueuedOut",
        "description": "Reference to a queued job."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}